Execute python files using python2 on all distros
On distros that default to python 3, the shebangs in src/generator.py
and test/mrdo-asm.py will cause the files to be executed with Python 3.
This version mismatch causes the scripts to crash. In the case of generator.py
no opcodes are generated and gb-disasm subsequently provides incorrect output.

Fix this by using /usr/bin/env to find the system's python2 binary